requirements.txt
flask==3.0.0
Pillow==10.2.0
werkzeug==3.0.1
requests==2.31.0
